+++ src/shared/Actions.js
@@ -29,7 +29,7 @@ export default {
 
     const validate = () =>
       this.debouncedValidation({
-        showErrors: this.state.options.get('showErrorsOnSubmit', this),
+        showErrors: this.state.options.get('showErrorsOnSubmit'),
         path: this.path,
       })
         .then(({ isValid }) => {